+++ src/mod.ts
@@ -2,4 +2,4 @@ export { YouTube } from "./youtube-deno.ts";
 export { YouTube_live } from "./live-streaming-api.ts";
 
 export * from "./oauth.ts";
-export * from "./schemas.ts";
+export * from "./schemas/schemas.ts";
